Perfeccionando métodos

+++ CafeteriApp/Cliente.cs
@@ -8,7 +8,7 @@ namespace CaferiApp
 {
     internal class Cliente : Usuario
     {
-        public Cliente(string permisos,string username, string contrasena,int telefono) : base(permisos,username, contrasena,telefono)
+        public Cliente(string permisos,string username, string contrasena, long telefono) : base(permisos,username, contrasena,telefono)
         {
         }
     }
